Updating glance-api-paste.ini for Grizzly support
This update removes all Keystone auth from the template, and
therefore does not need any variables passed into the erb template.
While we could simply remove this file altogether (in favor of this
being managed by the package manager du jour) as we would like to
support a pull from source model, best to leave it in and note that
it is managed by Chef.